feat(extract): emit scene metadata bundles
Summary:
- add scene metadata model types for actors, environment, paths,
  and exits
- extend SceneExtractor to emit metadata JSON alongside room
  geometry and verify those outputs
- expand display-list parsing for real Vendor/oot macros needed by
  the reviewer extraction gate

Rationale:
- TAN-67 requires extracting spot04 actor spawns, environment
  settings, paths, and exits from the real upstream scene sources
- the full-tree reviewer command was blocked by display-list macro
  gaps on current master, so the parser needed enough compatibility
  to let scene metadata extraction complete end to end

Tests:
- swift test --filter SceneExtractorTests
- swift test --filter DisplayListParserTests
- swift test
- swift run OOTExtractCLI extract --source /Users/tjohnson/repos/SwiftOOT/Vendor/oot --output /tmp/tan67-check
- swift run OOTExtractCLI verify --content /tmp/tan67-check

+++ Sources/OOTDataModel/F3DEX2Command.swift
@@ -261,9 +261,9 @@ public struct LoadUCodeExCommand: Codable, Sendable, Equatable {
 public struct BranchLessZCommand: Codable, Sendable, Equatable {
     public var branchAddress: UInt32
     public var vertexIndex: UInt16
-    public var zValue: UInt16
+    public var zValue: UInt32
 
-    public init(branchAddress: UInt32, vertexIndex: UInt16, zValue: UInt16) {
+    public init(branchAddress: UInt32, vertexIndex: UInt16, zValue: UInt32) {
         self.branchAddress = branchAddress
         self.vertexIndex = vertexIndex
         self.zValue = zValue
